char: store Chars as `char` instead of `u32`

+++ crates/flux-desugar/src/desugar.rs
@@ -1457,7 +1457,7 @@ trait DesugarCtxt<'genv, 'tcx: 'genv> {
             surface::LitKind::Bool => Ok(fhir::Lit::Bool(lit.symbol == kw::True)),
             surface::LitKind::Str => Ok(fhir::Lit::Str(lit.symbol)),
             surface::LitKind::Char => {
-                Ok(fhir::Lit::Char(u32::from(lit.symbol.as_str().parse::<char>().unwrap())))
+                Ok(fhir::Lit::Char(lit.symbol.as_str().parse::<char>().unwrap()))
             }
             _ => Err(self.emit_err(errors::UnexpectedLiteral { span })),
         }
