+++ crates/flux-refineck/src/checker.rs
@@ -1401,7 +1401,7 @@ impl<'ck, 'genv, 'tcx, M: Mode> Checker<'ck, 'genv, 'tcx, M> {
                 Ok(Ty::mk_ref(ReStatic, Ty::indexed(BaseTy::Str, idx), Mutability::Not))
             }
             Constant::Char(c) => {
-                let idx = Expr::constant(rty::Constant::from(*c as u128));
+                let idx = Expr::constant(rty::Constant::from(*c));
                 Ok(Ty::indexed(BaseTy::Char, idx))
             }
             Constant::Param(param_const, ty) => {
